fixup! Replaced math.h double functions by their float equivalent

+++ kernels/volk/volk_32f_asin_32f.h
@@ -133,7 +133,7 @@ volk_32f_asin_32f_a_sse4_1(float* bVector, const float* aVector, unsigned int nu
 
   number = quarterPoints * 4;
   for(;number < num_points; number++){
-    *bPtr++ = asin(*aPtr++);
+    *bPtr++ = asinf(*aPtr++);
   }
 }
 
@@ -200,7 +200,7 @@ volk_32f_asin_32f_u_sse4_1(float* bVector, const float* aVector, unsigned int nu
 
   number = quarterPoints * 4;
   for(;number < num_points; number++){
-    *bPtr++ = asin(*aPtr++);
+    *bPtr++ = asinf(*aPtr++);
   }
 }
 
@@ -216,7 +216,7 @@ volk_32f_asin_32f_u_generic(float* bVector, const float* aVector, unsigned int n
   unsigned int number = 0;
 
   for(number = 0; number < num_points; number++){
-    *bPtr++ = asin(*aPtr++);
+    *bPtr++ = asinf(*aPtr++);
   }
 }
 #endif /* LV_HAVE_GENERIC */
